feat: organize generated files in subfolders
