+++ brc100frames/abort-action.ts
@@ -1,4 +1,4 @@
-import {generateWireFramesFor} from "./wire-frames-generator";
+import {generateWireFramesFor} from "./generator/wire-frames-generator";
 
 
 export const abortAction: Record<string, ReturnType<typeof generateWireFramesFor>> = {
