+++ brc100frames/create-action.ts
@@ -1,4 +1,4 @@
-import {generateWireFramesFor} from "./wire-frames-generator";
+import {generateWireFramesFor} from "./generator/wire-frames-generator";
 import { Utils } from '@bsv/sdk';
