+++ brc100frames/generator/wire-frames-generator.ts
